# create-a-container/openapi.v1.yaml
openapi: 3.0.3
info:
  title: Create-a-Container API v1
  description: |
    JSON-only REST API for the React SPA and external automation.
    Co-exists with the legacy `/` (HTML) and `/apikeys` (mixed) routes during the React migration.

    ## Authentication

    Two mechanisms are supported:

    - **Session cookie** — established by `POST /api/v1/auth/login`. Requires a CSRF token (see below) on every state-changing request.
    - **API key** — `Authorization: Bearer <key>`. CSRF is skipped for Bearer-authenticated requests.

    ## CSRF (session auth only)

    1. `GET /api/v1/csrf-token` → returns `{ "data": { "csrfToken": "..." } }` and sets `__Host-csrf.token` cookie.
    2. Send the token in the `X-CSRF-Token` header on every `POST`/`PUT`/`PATCH`/`DELETE`.

    ## Response envelope

    Success: `{ "data": <payload>, "meta"?: {...} }`.
    Error:   `{ "error": { "code": "...", "message": "...", "fields"?: {...} } }`.
  version: 1.0.0
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    SessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
...
security:
  - BearerAuth: []
  - SessionCookie: []
paths:
...
  /sites/{siteId}/containers:
    get:
      tags: [Containers]
      parameters:
        - { in: path, name: siteId, required: true, schema: { type: integer } }
        - { in: query, name: hostname, schema: { type: string } }
      responses: { '200': { description: Array of containers } }
    post:
      tags: [Containers]
      parameters: [{ in: path, name: siteId, required: true, schema: { type: integer } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [hostname, template]
              properties:
                hostname: { type: string }
                template: { type: string }
                customTemplate: { type: string }
                entrypoint: { type: string }
                nvidiaRequested: { type: boolean }
                environmentVars:
                  type: array
                  items: { type: object, properties: { key: { type: string }, value: { type: string } } }
                services:
                  type: object
                  additionalProperties: { type: object }
      responses: { '201': { description: Creation job enqueued } }
...
  /sites/{siteId}/containers/{id}:
    parameters:
      - { in: path, name: siteId, required: true, schema: { type: integer } }
      - { in: path, name: id, required: true, schema: { type: integer } }
    get: { tags: [Containers], responses: { '200': { description: Container } } }
    put: { tags: [Containers], responses: { '200': { description: Updated, optional restart job } } }
    delete: { tags: [Containers], responses: { '200': { description: Deleted, with DNS cleanup warnings } } }
...